+++ resources/js/utilities/api/resources/activity.js
@@ -13,4 +13,8 @@ export default class extends BaseResource{
     progress(activityId, page=1, perPage=10) {
         return this.request('get', '/activity/' + activityId + '/progress', {}, {page: page, per_page: perPage});
     }
+
+    delete(activityId) {
+        return this.request('delete', '/activity/' + activityId);
+    }
 }