+++ resources/js/utilities/api/resources/moduleinstances.js
@@ -10,4 +10,7 @@ export default class extends BaseResource{
         return this.request('patch', '/module-instance/' + id, attributes);
     }
 
+    delete(id) {
+        return this.request('delete', 'module-instance/' + id);
+    }
 }