+++ resources/views/errors/error.blade.php
@@ -7,7 +7,7 @@
         message="{{($exception instanceof \Symfony\Component\HttpKernel\Exception\HttpException) ? $exception->getMessage() : ''}}"
         code="{{method_exists($exception, 'getStatusCode') ? $exception->getStatusCode() : $exception->getCode()}}"
         email="bristol-su@bristol.ac.uk"
-        home-url="{{route(request()->is(['a', 'a/*']) ? 'admin' : 'participant')}}"
+        home-url="{{route(request()->is(['a', 'a/*']) ? 'administrator' : 'participant')}}"
     >
 
     </p-error-page>